+++ Final/src/Model/MathFunctions.java
@@ -99,7 +99,7 @@ public class MathFunctions
 			throw new SyntaxErrorException("Factorial only defined for integers");
 		if(n < 0)
 			throw new IllegalArgumentException("No negative values allowed");
-		int fact = 1;
+		double fact = 1;
 		for (int i = 2; i <= n; i++)
 			fact = fact * i;
 		return fact;
@@ -896,3 +896,4 @@ public class MathFunctions
 
 	}
 }
+
